box ParseError variants to eliminate 310+ clippy result_large_err warnings
- Box Option<ParseContext> in all 13 ParseError variants (~72 bytes saved each)
- Extract ContextualSyntaxError to Box<ContextualSyntaxErrorData> (208 -> 8 bytes)
- Update convenience constructors to box internally (callers unchanged)
- Fix borrow-after-move in macro expansion.rs
- Parser crate: 0 result_large_err warnings (was ~310)

+++ src/compiler_rust/compiler/src/arch_rules.rs
@@ -193,7 +193,7 @@ impl ArchRulesChecker {
         // Extract type usage from functions
         for func in &module.functions {
             // Get line number from function span if available
-            let func_line = func.span.map(|s| s.line as usize).unwrap_or(0);
+            let func_line = func.span.map(|s| s.line).unwrap_or(0);
 
             // Extract from function signature (parameters and return type)
             for param in &func.params {
